classifier.visual method
Pass feature and target into the method to see the 1d and 2d density function. 
Should always pass the target variable when wanting to visualize in 2d since predicting the class value depends on estimating the join probabilities of all attribute and target variables. Thus, a 2d visualization is theoretically a visualization of joint probability?