H2 embedded DB
When executing the classifier.train() method, the syntax prints that the quantization table (_qt) has been dropped and successfully created (by printing “ok”). Though, this is in fact not the case and is problematic because table _qmt is based off _qt and so on so forth. I’ve troubleshooted it by instead using the executeQuery function on each individual jinja template for _qt, _qmt, and _m by hard coding the associated variables. This approach seems to work fine, and all tables are successfully and actually created in the H2 DB. But, this obviously is not what we want, and I am currently out of ideas for why the materializedView  says the table has been created when in actuality, it has not.