Update database.py
I am adding explicitly to the mysql connection to authenticate on username and password on the database. This method is by default. However by compiling these libraries on ARM architectures like M1 Chips, this is not the case.

In order to extend compatibility with M1 or ARM chips. It is important this change